pyproject.toml
[project]
name = "pulumi-dagger-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anyio>=4.9.0",
    "dagger-io>=0.18.6",
    "pulumi>=3.169.0",
    "pulumi-aws>=6.80.0",
    "python-dotenv>=1.1.0",
]
